Refactor DTO package and add transaction status field
Moved DTO classes to the controller.dto package for better structure and alignment. Added a new `TransactionStatus` field to `TransactionFilterRequest` to support filtering by transaction status. This improves code organization and extends functionality for transaction filtering.
