+++ src/main/java/info/mackiewicz/bankapp/system/banking/history/controller/TransactionHistoryRestController.java
@@ -1,6 +1,6 @@
 package info.mackiewicz.bankapp.system.banking.history.controller;
 
-import info.mackiewicz.bankapp.system.banking.history.dto.TransactionFilterRequest;
+import info.mackiewicz.bankapp.system.banking.history.controller.dto.TransactionFilterRequest;
 import info.mackiewicz.bankapp.system.banking.history.service.TransactionHistoryService;
 import info.mackiewicz.bankapp.system.banking.shared.dto.TransactionResponse;
 import info.mackiewicz.bankapp.user.model.User;
